+++ internal/service/llm/fcall/fcall_ask_user.go
@@ -19,8 +19,5 @@ func (a *AskUserFunctionCall) Call(ctx *Context, req Request) (Response, error) 
 		return Response{}, err
 	}
 	ctx.SetAttachment(a.Name(), val)
-	return Response{
-		Output: "请稍等，用户的回答将在下一次请求中.....",
-		Status: "in_progress",
-	}, nil
+	return Response{}, nil
 }
