+++ internal/service/llm/platform/openai/handler.go
@@ -61,16 +61,7 @@ func (h *Handler) stream(ctx context.Context, req domain.StreamRequest) (stream 
 }
 
 func (h *Handler) newParams(req domain.StreamRequest) responses.ResponseNewParams {
-	// messages := h.getMessages(req)
-	// for i := range messages {
-	// 	log.Printf("newParams messages[%d]: %#v\n", i, messages[i])
-	// }
-	for i := range req.Messages {
-		log.Printf("req messages[%d]: %#v\n", i, req.Messages[i])
-	}
-
 	params := responses.ResponseNewParams{
-		// Input: h.toInput(messages),
 		Input: h.toInput(req.Messages),
 		Model: req.ConfigVersion.Model.Name,
 		Tools: slice.Map(req.ConfigVersion.Functions, func(_ int, src domain.Function) responses.ToolUnionParam {
@@ -102,43 +93,6 @@ func (h *Handler) newParams(req domain.StreamRequest) responses.ResponseNewParam
 		params.Instructions = openai.String(req.ConfigVersion.SystemPrompt)
 	}
 	return params
-}
-
-func (h *Handler) getMessages(req domain.StreamRequest) []domain.Message {
-	systemContent := ""
-	userContent := ""
-	if req.ConfigVersion.SystemPrompt != "" {
-		systemContent += req.ConfigVersion.SystemPrompt
-		systemContent += "\n"
-	}
-	if req.ConfigVersion.Prompt != "" {
-		userContent += req.ConfigVersion.Prompt
-		userContent += "\n"
-	}
-	for _, message := range req.Messages {
-		if message.Role == domain.SYSTEM && message.Content != "" {
-			systemContent += message.Content
-			systemContent += "\n"
-		}
-		if message.Role == domain.USER && message.Content != "" {
-			userContent += message.Content
-			userContent += "\n"
-		}
-	}
-	messages := make([]domain.Message, 0, 2)
-	if systemContent != "" {
-		messages = append(messages, domain.Message{
-			Role:    domain.SYSTEM,
-			Content: systemContent,
-		})
-	}
-	if userContent != "" {
-		messages = append(messages, domain.Message{
-			Role:    domain.USER,
-			Content: userContent,
-		})
-	}
-	return messages
 }
 
 func (h *Handler) toInput(messages []domain.Message) responses.ResponseNewParamsInputUnion {
@@ -173,66 +127,37 @@ func (h *Handler) toInput(messages []domain.Message) responses.ResponseNewParams
 
 func (h *Handler) forward(req domain.StreamRequest, stream *ssestream.Stream[responses.ResponseStreamEventUnion], events chan domain.StreamEvent) {
 	// defer close(events)
+	for i := range req.Messages {
+		log.Printf("req.messages[%d]: %#v\n", i, req.Messages[i])
+	}
 	fCtx := &fcall.Context{Context: context.Background()}
-	funcCallOutputs := make([]responses.ResponseInputItemUnionParam, 0, 10)
+	var hasLLMInvokeCall bool
 	var responseID string
 	for stream.Next() {
-
 		event := stream.Current()
 		switch event.Type {
 		case "error":
 			events <- domain.StreamEvent{
-				ResponseID: responseID,
-				Error:      errors.New(event.AsError().Message),
-				Done:       true,
+				Error: errors.New(event.AsError().Message),
+				Done:  true,
 			}
 		case "response.created":
 			responseID = event.Response.ID
 			log.Printf("forward: response created ID = %s\n", responseID)
 		case "response.completed":
-			// if len(funcCallOutputs) > 0 {
-			// 	log.Printf("有函数调用需要发送响应，暂不返回\n")
-			// 	req.Messages = nil
-			// 	req.PreviousResponseID = responseID
-			// 	params := h.newParams(req)
-			// 	params.Input.OfInputItemList = funcCallOutputs
-			// 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
-			// 	h.client.Responses.NewStreaming(ctx, params)
-			// 	cancel()
-			// } else {
-			// 	evt := domain.StreamEvent{
-			// 		ResponseID: responseID,
-			// 		Done:       true,
-			// 	}
-			// 	events <- evt
-			// 	log.Printf("没有函数调用需要发送响应，直接返回: event=%#v\n", evt)
-			// }
-
-			noFuncCalls := len(funcCallOutputs) == 0
-			if noFuncCalls {
-				evt := domain.StreamEvent{
-					ResponseID: responseID,
-					Done:       true,
+			if !hasLLMInvokeCall {
+				events <- domain.StreamEvent{
+					Done: true,
 				}
-				events <- evt
-				log.Printf("没有函数调用需要发送响应，直接返回: event=%#v\n", evt)
-			} else {
-				log.Printf("有函数调用需要发送响应，暂不返回\n")
 			}
-
 		case "response.output_text.delta":
 			text := event.AsResponseOutputTextDelta()
 			events <- domain.StreamEvent{
-				ResponseID: responseID,
-				Content:    text.Delta,
+				Content: text.Delta,
 			}
 		case "response.reasoning_summary_text.delta":
 			r := event.AsResponseReasoningSummaryTextDelta()
 			events <- domain.StreamEvent{
-				ResponseID: responseID,
-				// ReasoningContent: map[int64]string{
-				// 	r.SummaryIndex: r.Delta,
-				// },
 				ReasoningContent: r.Delta,
 			}
 		case "response.output_item.done":
@@ -252,61 +177,50 @@ func (h *Handler) forward(req domain.StreamRequest, stream *ssestream.Stream[res
 				continue
 			}
 
-			var output responses.ResponseInputItemUnionParam
-			resp, err := fn.Call(fCtx, fcall.Request{Args: []byte(fc.Arguments)})
+			_, err = fn.Call(fCtx, fcall.Request{Args: []byte(fc.Arguments)})
 			if err != nil {
 				h.logger.Error("执行函数调用失败",
 					elog.FieldCustomKeyValue("函数名", fc.Name),
 					elog.FieldCustomKeyValue("参数值", fc.Arguments),
 					elog.FieldErr(err),
 				)
-				output = responses.ResponseInputItemParamOfFunctionCallOutput(fc.CallID, err.Error())
-			} else {
-				output = responses.ResponseInputItemUnionParam{
-					OfFunctionCallOutput: &responses.ResponseInputItemFunctionCallOutputParam{
-						CallID: fc.CallID,
-						Output: resp.Output,
-						Status: resp.Status,
-					},
-				}
 			}
-			if fn.Name() == fcall.NameAskUser {
+			switch fn.Name() {
+			case fcall.NameAskUser:
 				events <- domain.StreamEvent{
-					ResponseID: responseID,
-					CallID:     fc.CallID,
-					Content:    fCtx.GetAttachment(fcall.NameAskUser),
+					Content: fCtx.GetAttachment(fcall.NameAskUser),
 				}
-			} else {
-				funcCallOutputs = append(funcCallOutputs, output)
+			case fcall.NameEmitJSON:
+				log.Printf("emit_json: 提取到的结构化数据: %#v\n", fCtx.JSONData)
+			case fcall.NameInvokeLLM:
+				hasLLMInvokeCall = true
+				var msg domain.Message
+				_ = json.Unmarshal([]byte(fCtx.GetAttachment(fcall.NameInvokeLLM)), &msg)
+				log.Printf("invok_llm: %#v", msg)
+				req.Messages = append(req.Messages, msg)
+				params := h.newParams(req)
+				// params.PreviousResponseID = openai.String(responseID)
+				// params.Input.OfInputItemList = append(params.Input.OfInputItemList, responses.ResponseInputItemUnionParam{
+				// 	OfFunctionCallOutput: &responses.ResponseInputItemFunctionCallOutputParam{
+				// 		CallID: fc.CallID,
+				// 		Output: msg.Content,
+				// 		Status: "completed",
+				// 	},
+				// })
+				ctx, cancel := context.WithTimeout(context.Background(), time.Second*1000)
+				h.forward(req, h.client.Responses.NewStreaming(ctx, params), events)
+				cancel()
 			}
 		}
-
 	}
+
 	err := stream.Err()
 	if err != nil {
 		h.logger.Error("forward过程中出错", elog.FieldErr(err))
 		log.Printf("forward过程中出错: %+v\n", err)
-	}
-
-	if len(funcCallOutputs) > 0 {
-		req.Messages = nil
-		req.PreviousResponseID = responseID
-		params := h.newParams(req)
-		ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
-		h.forward(req, h.client.Responses.NewStreaming(ctx, params), events)
-		cancel()
-	}
-}
-
-func (h *Handler) newInputItem(text string) responses.ResponseInputItemUnionParam {
-	return responses.ResponseInputItemUnionParam{
-		OfInputMessage: &responses.ResponseInputItemMessageParam{
-			Content: []responses.ResponseInputContentUnionParam{{
-				OfInputText: &responses.ResponseInputTextParam{
-					Text: text,
-				},
-			}},
-			Role: "user",
-		},
+		// events <- domain.StreamEvent{
+		// 	Error: err,
+		// 	Done:  true,
+		// }
 	}
 }
