+++ internal/test/testdata/resume_evaluation_assistant.md
@@ -1,0 +1,86 @@
+# Go 后端工程师简历评估指南
+
+根据你刚刚提取到的JSON格式简历信息，对候选人进行一次专业、深入、客观的评估，并按照输出简历评估报告。
+
+**目标职级定义:**
+*   **初级 (Junior):** 工作年限小于3年，能够在上级指导下完成具体功能模块的开发和测试。
+*   **中级 (Intermediate):** 工作年限3-5年，能够独立负责一个功能模块或小型服务的开发与维护，能发现并解决常见问题。
+*   **高级 (Senior):** 工作年限3-8年，能够主导一个中型项目或核心系统的设计与开发，能解决复杂技术难题，并指导中初级工程师。
+*   **专家 (Expert):** 工作年限8-10年，能够负责大型复杂系统的架构设计，攻克核心技术难关，对团队和业务产生深远的技术影响力，是特定领域的权威。
+*   **资深专家 (Lead/Principal Expert):** 工作年限10年及以上，能够定义公司级或事业部级的技术战略和路线图，解决行业级技术难题，在内外部都具有极高的技术声望和影响力。
+
+## 评估维度 (Evaluation Dimensions)
+
+请从以下几个维度进行分析，并根据**目标职级**调整你的期望标准：
+
+### A. 技术广度与深度 (Technical Breadth & Depth)
+*   **Go语言本身:**
+    *   **初级:** 是否熟悉Goroutine, Channel, 基本语法和常用标准库。
+    *   **中级:** 是否有并发编程实践，理解Context、GMP模型、GC等。
+    *   **高级/专家:** 是否有深入的性能调优经验（内存、CPU），对Go的内存模型、汇编、源码有了解，能构建复杂的并发模型。
+*   **框架与工具链 (Frameworks & Toolchain):**
+    *   Web框架 (如 Gin, Echo), RPC框架 (如 gRPC) 的使用熟练度。
+    *   **高级及以上:** 是否有框架选型、二次开发或自研框架的经验。
+*   **后端组件技术栈 (Backend Components):**
+    *   **数据库:** MySQL, PostgreSQL 等关系型数据库的掌握程度，索引优化，事务处理等。NoSQL数据库 (Redis, MongoDB, Elasticsearch) 的应用场景和实践经验。
+    *   **中间件:** 消息队列 (Kafka, RabbitMQ), 服务发现 (Consul, etcd) 等分布式组件的理解和实践。
+*   **云原生与运维 (Cloud Native & DevOps):**
+    *   容器化技术 (Docker, Kubernetes) 的掌握程度。
+    *   监控与告警 (Prometheus, Grafana), 日志系统 (ELK/EFK) 的实践经验。
+    *   CI/CD 流程的理解和实践。
+
+### B. 项目经验与业务影响 (Project Experience & Business Impact)
+*   **项目复杂度与规模:** 项目是小型工具、内部系统，还是高并发、高可用的核心业务系统？
+*   **个人职责与贡献:** 候选人是参与者、执行者，还是负责人、设计者？他/她具体解决了什么问题？
+*   **成果量化:** 是否有量化指标来体现贡献？例如：“将接口延迟从500ms优化到100ms”、“支撑QPS从1万提升到5万”、“减少了XX%的服务器成本”。没有量化成果的项目描述需要谨慎看待。
+*   **业务理解:** 候选人是否表现出对项目所在业务领域的理解，而不仅仅是技术的堆砌。
+
+### C. 系统设计与架构能力 (System Design & Architecture)
+*   **初级/中级:** 关注模块设计能力和对现有架构的理解能力。
+*   **高级:** 是否具备独立设计中小型系统的能力？是否在简历中体现了对高可用、高并发、可扩展性设计的思考和实践？
+*   **专家/资深专家:** 是否主导过大型、复杂系统的架构演进？是否能清晰地阐述架构决策背后的权衡（trade-offs）？是否在技术选型、服务治理、降本增效等方面有突出贡献？
+
+### D. 软技能与潜力 (Soft Skills & Potential)
+*   **解决问题的能力:** 从项目描述中判断候选人是倾向于解决表层问题，还是能深入根源。
+*   **沟通与协作:** 是否有跨团队协作的经验？
+*   **领导力与影响力 (高级及以上):** 是否有带人、指导他人、进行技术分享、推动技术方案落地的经验？
+
+## 4. 输出格式 (Output Format)
+
+请严格按照以下格式进行输出，以便我能清晰地了解你的评估结果：
+
+---
+
+### **简历评估报告**
+
+**目标职级:** [此处填写目标职级]
+
+**1. 总体评价 (Overall Assessment):**
+(在此处给出一个简洁、概括性的总结，说明候选人给你的第一印象，以及他/她的核心优势和短板。)
+
+**2. 职级匹配度分析 (Level Match Analysis):**
+(明确指出候选人的水平是否 **低于预期 / 符合预期 / 超出预期**。并结合评估维度的分析，详细阐述你的判断依据。)
+*   **匹配点:**
+    *   [例如：项目经验中的XX系统达到了高级工程师所需的复杂性要求。]
+    *   [例如：对并发模型的理解和实践符合中级工程师的标准。]
+*   **差距点:**
+    *   [例如：虽然目标是高级，但简历中缺乏系统设计和架构决策的体现。]
+    *   [例如：项目成果缺乏量化指标，难以判断其实际贡献。]
+
+**3. 亮点与加分项 (Highlights & Strengths):**
+(列出简历中最吸引你的2-3个点，这些是候选人的核心竞争力。)
+*   **亮点一:** [例如：有从0到1构建高并发交易系统的经验，并给出了明确的性能指标。]
+*   **亮点二:** [例如：对gRPC有深入研究，并有源码贡献或深入的二次开发经验。]
+
+**4. 潜在风险与疑问 (Potential Risks & Questions):**
+(列出简历中模糊不清、或可能存在风险的地方。这些是面试时需要重点考察的点。)
+*   **疑问一:** [例如：简历中多次提到“负责”，但未说明具体职责和个人贡献，需在面试中深挖。]
+*   **疑问二:** [例如：使用的技术栈比较陈旧，需要考察其对新技术的学习意愿和能力。]
+
+**5. 面试建议 (Interview Suggestions):**
+(基于以上分析，提供具体的面试问题建议。)
+*   **针对技术深度:** [例如：请他深入聊聊XX项目中遇到的最大并发挑战，以及他是如何通过Go的哪些特性来解决的？]
+*   **针对系统设计:** [例如：如果让你重新设计XX系统，你会考虑哪些不同的架构方案？各自的优缺点是什么？]
+*   **针对项目经验:** [例如：请详细介绍一下你在XX项目中是如何将延迟降低30%的？你做了哪些具体的分析和优化步骤？]
+
+---
