+++ internal/test/testdata/resume_xiaoming_missing_info.md
@@ -1,0 +1,51 @@
+# 个人简历
+
+## 基本信息
+- **姓名**: 李小明
+- **现居地**: 北京市朝阳区
+
+## 教育背景
+- **毕业院校**: 北京理工大学
+- **学历**: 本科
+- **专业**: 计算机科学与技术
+- **毕业时间**: 2021年6月
+
+## 求职意向
+- **目标岗位**: Go后端工程师
+
+
+## 专业技能
+- **编程语言**: Go、Python、JavaScript
+- **数据库**: MySQL、Redis、MongoDB
+- **框架技术**: Gin、Echo、Vue.js
+- **开发工具**: Docker、Kubernetes、Git
+- **云服务**: 阿里云、AWS基础服务
+
+## 工作经历
+
+### 北京创新科技有限公司 (2021.07 - 至今)
+**职位**: 初级Go后端工程师
+
+**主要职责**:
+- 负责公司核心业务系统的后端开发和维护
+- 参与微服务架构设计，使用Go语言开发RESTful API
+- 负责数据库设计和优化，主要使用MySQL和Redis
+- 配合前端团队完成接口对接和联调测试
+
+**项目经验**:
+1. **用户管理系统** (2021.08-2021.12)
+    - 技术栈：Go + Gin + MySQL + Redis
+    - 负责用户注册、登录、权限管理等核心功能开发
+
+2. **订单处理系统** (2022.01-2022.08)
+    - 技术栈：Go + gRPC + MongoDB
+    - 实现了高并发订单处理逻辑，日处理订单量10万+
+
+## 个人评价
+作为一名初级Go后端工程师，我具备扎实的编程基础和良好的学习能力。在2年的工作经验中，我参与了多个核心项目的开发，熟练掌握Go语言生态及相关技术栈。我注重代码质量，具有良好的团队协作精神，能够快速适应新的技术和业务需求。
+
+## 自我评价
+- 学习能力强，对新技术保持敏感和热情
+- 代码规范，注重系统性能和可维护性
+- 具备良好的问题分析和解决能力
+- 团队协作意识强，沟通能力良好
